⏰ Refactors ntp service
